app/alg/ClassMood/audio-service/requirements.txt
Flask==2.3.3
ffmpeg-python==0.2.0
librosa==0.10.1
scipy==1.11.2
numpy>=1.21